requirements.txt
# YOLOv7 訓練系統依賴套件
# Python 3.13.9 相容版本

# 核心深度學習框架
torch>=2.0.0
torchvision>=0.15.0

# 數值計算與影像處理
numpy>=1.24.0
opencv-python>=4.8.0

# 設定檔解析
PyYAML>=6.0

# 訓練進度顯示
tqdm>=4.65.0

# 視覺化工具
matplotlib>=3.7.0

# 影像處理輔助
Pillow>=10.0.0

# 科學計算 (用於 mAP 計算)
scipy>=1.11.0

# 資料處理
pandas>=2.0.0
